pyproject.toml
[project]
name = "st-api"
version = "0.1.0"
description = "Shop Titans text classification API for extracting digits from images"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "modal>=1.0.2",
  "fastapi[standard]>=0.100.0",
  "numpy>=2.2.4",
  "pillow>=11.2.1",
  "scikit-learn>=1.6.1",
  "python-multipart>=0.0.6",
]
